Modified ```np.int32``` to ```int``` to resolve AttributeError thrown by Numpy>=1.20.0 (#253)
* modified np.int32 to int to resolve AttributeError thrown by Numpy>=1.20.0

* removed comment from pretty_midi.py, updated np.int32 aliase in instruments.py

+++ pretty_midi/pretty_midi.py
@@ -346,7 +346,7 @@ class PrettyMIDI(object):
             last_note_on = collections.defaultdict(list)
             # Keep track of which instrument is playing in each channel
             # initialize to program 0 for all channels
-            current_instrument = np.zeros(16, dtype=np.int32)
+            current_instrument = np.zeros(16, dtype=int)
             for event in track:
                 # Look for track name events
                 if event.type == 'track_name':
